agent-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
pymysql==1.1.0
openai>=1.0.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy>=1.26.0
scikit-learn==1.3.2
schedule==1.2.0
websockets==12.0
python-multipart==0.0.6